requirements.txt
markdownify
smolagents
requests
duckduckgo_search
pandas
transformers
# transformers[agents]
torch
langchain
openai
# accelerate
langchain-community
google-search-results
smolagents[transformers]
scikit-image
pycountry
opentelemetry-sdk
opentelemetry-exporter-otlp
openinference-instrumentation-smolagents
python-dotenv
langchain-openai